\name{Dilution}
\docType{data}
\alias{Dilution}
\title{AffyBatch instance Dilution}
\description{
This \code{\link[affy]{AffyBatch-class}} object represents part of
a dilution experiment dataset. 
}
\usage{data(Dilution)}
\format{
  An \code{\link[affy]{AffyBatch-class}} object containing 4 arrays.
}
\source{Two sources of cRNA A (human liver tissue) and B (Central Nervous
System cell line) have been hybridized to human array (HGU95A) in a range
of proportions and dilutions. This data set is taken from
arrays hybridized to source A at 10.0 and 20 \eqn{\mu}g. We have two
replicate arrays for each generated cRNA. Three scanners have been used
in this study. Each array replicate was processed in a different scanner.

For more information see Gautier et al., affy - Analysis of Affymetrix
GeneChip data at the probe level
\url{http://bioinformatics.oxfordjournals.org/content/20/3/307.full.pdf}
Bioinformatics, 2004
}
\keyword{datasets}
